Update black pawn movegen.
This commit removes move generation for black pawns on the first and
last ranks of the board. No pawn can move backwards, and any pawn that
reaches the last rank is promoted, so this is fine. This means the Move
Table will return `Option::None` if the program attempts to look up pawn
moves on these squares. There should be no case in which this happens,
as that would indicate a severe logic bug.