Everything changed.
This overhauls the way MoveTable is used by boxing it in a new struct
NoArc. NoArc is a thread-safe way to read a block of memory. Essentially
every file in the engine had to change, because essentially every file
used GameManager or MoveTable in some way. It shouldn't be too horrible
to sort through, though!